requirements.txt
oracledb
requests
pandas
python-dotenv